Actualizar el sistema: KalmanUpdater

El tipo abstracto KalmanUpdater está pensado como una interfaz a las estructuras que permitan actualizar el sistema desde un estado $x_{n}$ a un estado $x_{n+1}$.

Para que funciones correctamente deben implementarse los siguientes métodos:

Métodos a implementar	Breve descripción

update!(L::KalmanUpdater, hatx, hatP, control, t)	un método que permita actualizar al iterador y dejarlo listo para la siguiente iteración. Cuando se usan matrices $M_n := M, B_n := B, F_n:= F$ contantes se puede dejar en blanco, pero debería usarse, por ejemplo, para linearlizar en torno a $\hat{x}_n$ cuando se usa un KalmanUpdater no lineal.

forecast(updater::KalmanUpdater, hatx, hatP, control, t)	Devuelve una tupla que contiene a $\hat{x}_{n+1, n}, \hat{P}_{n+1, n}$ a partir de $\hat{x}_{n,n}$(hatx), $\hat{P}_{n,n}$(hatP) y un control.

dt(updater::KalmanUpdater)	$\Delta t$

LinearizableUpdater

Un grupo importante de KalmanUpdaters serán los que, o bien son lineales de la forma

\[x_{n+1} = M_n x_n + B_n u_n + F_n N_n\]

(donde $M_n, F_n$ son matrices, $x_n, N_n, B_n$ vectores y $u_n$ es un control escalar), o bien, en caso de ser no lineales $x_{n+1} = \mathcal{M}(x_n, u_n)$, pueden ser linealizados a la forma anterior.

Los miembros de esta clase necesitan implementar:

Métodos a implementar	Breve descripción

Mn, Bn, Fn	De la linearización en el estado actual

Los miembros de esta clase tienen automáticamente definidos los métodos:

forecast(updater::LinearizableUpdater, hatx, hatP, control, t)
forecast_with_error(updater::LinearizableUpdater, hatx, hatP, control, t)
forecast_hatP(updater::LinearizableUpdater, hatP)

Forecast con error

Para usar EnKF<!–TODO: agregar referencia a EnKF–> es necesario que el KalmanUpdater usado tenga implementado forecast_with_error. Esto permite hacer un paso de forecast e incluir error.

Cuentan además con las siguientes funciones

KalmanFilter.dimensions — Function

dimensions(updater::KalmanFilter.LinearizableUpdater) -> Any

Devuelve el número de dimensiones que puede soportar el LinearizableUpdater en el estado.

dimensions(observer::KalmanFilter.KalmanObserver) -> Any

Devuelve el número de dimensiones que puede soportar el LinearizableUpdater en el estado.

KalmanFilter.noiser — Function

noiser(updater::KalmanFilter.LinearizableUpdater) -> Any

Devuelve una distribución normal multivariada $\mathcal{N}(0, Q)$, donde $Q$ es la matriz de covarianzas del LinearizableUpdater.

los KalmanUpdaters de este tipo implementados son:

SimpleLinearUpdater

KalmanFilter.SimpleLinearUpdater — Type

struct SimpleLinearUpdater{T} <: KalmanFilter.LinearizableUpdater

Define un actualizador lineal constante.

Define un actualizador lineal del sistema, que permite calcular el estado $x_{n+1}$ a partir del estado $x_{n}$ y un control escalar $u_n$ según

\[x_{n+1} = M x_n + B u_n + F N_n\]

donde $N_n$ es un número aleatorio (dado por una variable aleatorio normal $\mathcal{N}(0,1)$).

Campos

M::AbstractArray{T,2} where T

Matriz $M$

B::AbstractArray{T,1} where T

Vector $B$

F::AbstractArray{T,2} where T

Matriz $F$

Q::AbstractArray{T,2} where T

Matriz $Q$ de covarianzas del error

dt::Any

$\Delta t$

integrity::Any

Función que corrige x para dejarlo dentro de un dominio.

NLUpdater

KalmanFilter.NLUpdater — Type

mutable struct NLUpdater <: KalmanFilter.LinearizableUpdater

Define un actualizador dado por la discretización de una ODE.

Define un actualizador posiblemente no lineal del sistema, a partir de la discretización de una Ecuación Diferencial Ordinaria. Esto permite calcular el estado $x_{n+1}$ a partir del estado $x_{n}$ y el control $u_n$ según

\[x_{n+1} = \mathcal{M}(x_n, u_n) + F N_n\]

donde $N_n$ es un número aleatorio (dado por una variable aleatorio normal $\mathcal{N}(0,Q)$) y $\mathcal{M}$ está dada por un Discretizer.

En todo tiempo guarda una versión linealizada del sistema, en torno a un punto $x_n, u_n, t_n$ (al construirlo se le deben dar esos valores también.)

Ambas estructuras pueden ser además evaluadas en la firma (::KalmanUpdater)(x::AbstractArray, u::Real, error). Por ejemplo, para el caso no lineal, nlupdater(x,u,ε) devuelve $\mathcal{M}(x,u) + Fε$.

Otros KalmanUpdaters implementados

Un actualizador importante es ODEForecaster, que ofrece un enfoque alternativo para trabajar con ecuaciones diferenciales estocásticas de la forma

\[x(t) = F(x, t)dt + G(x, t)dW(t)\]

Este método se llama Filtro de Kalman Continuo Discreto, y es muy similar al Filtro de Kalman Extendido, pero el paso de forecast se hace resolviendo una ecuación diferencial acoplada, la Ecuación de los Momentos, para $\hat{x}$ y $\hat{P}$:

\[\begin{aligned} \hat{x}'(t) &= F(\hat{x},t) \\ \hat{P}'(t) &= \partial_\hat{x} F(\hat{x}, t) \hat{P} + \hat{P} \left( \partial_\hat{x} F(\hat{x}, t) \right) ^{\top} + G(t)Q(t)G(t)' \end{aligned}\]

De esta forma, para obtener $\hat{x}_{n+1,n}, \hat{P}_{n+1,n}$ basta con resolver la ecuación de los momentos en $[t_{n}, t_{n+1}]$, con condiciones iniciales $\hat{x}_{n+1,n}, \hat{P}_{n+1,n}$.

ODEForecaster

KalmanFilter.ODEForecaster — Type

Outer constructor

Consideraciones adicionales

Función de integridad

En general los constructores pedirán además una función de integridad; cuya idea es evitar que los estados se salgan de cierto dominio. Por ejemplo, para el caso en que trabajamos con EDOs epidemiológicas, es necesario que las cantidades sean no negativas, por lo que se usa la función de integridad (x) -> max.(x, 0.), que transforma un vector x para dejar en 0 las coordenadas negativas.

update_inner_system y update_aproximation

Aún no está definida del todo la interfaz de KalmanUpdater, y varias estructuras implementan estos dos métodos.

CREO que ya ni siquiera es necesaria la distinción, pues venía del hecho de que KalmanIterator manejaba los ruidos, cosa que ahora hace KalmanUpdater. En todo caso, ciertos actualizadores como ODEForecaster no tienen definido un método update_inner_system porque no están pensados para hacer eso (solo deberían usarse con Measurements, no con InnerStates). Es necesario implementar holy traits o algo así para diferenciar mejor los métodos que deben implementarse.